% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileProvider.R
\name{AbstractFileProvider}
\alias{AbstractFileProvider}
\title{File providers}
\description{
File providers

File providers
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AbstractFileProvider$new()}}
\item \href{#method-listFiles}{\code{AbstractFileProvider$listFiles()}}
\item \href{#method-listAllFiles}{\code{AbstractFileProvider$listAllFiles()}}
\item \href{#method-relativePath}{\code{AbstractFileProvider$relativePath()}}
\item \href{#method-print}{\code{AbstractFileProvider$print()}}
\item \href{#method-clone}{\code{AbstractFileProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
New file provider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFileProvider$new(sep = .Platform$file.sep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{- platform separator}

\item{\code{...}}{for compatibility}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the provider
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listFiles"></a>}}
\subsection{Method \code{listFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFileProvider$listFiles(directory = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{directory}}{- the directory}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a dataframe containing the full path relative to root of this provider (path) and file / directory flag (isDir)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listAllFiles"></a>}}
\subsection{Method \code{listAllFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFileProvider$listAllFiles(directory = ".")}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a dataframe containing the full path relative to root of this provider (path) of files and directories and file / directory flag (isDir)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-relativePath"></a>}}
\subsection{Method \code{relativePath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFileProvider$relativePath(parts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFileProvider$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFileProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
