% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawardtools.R
\name{applyAgeTimeMatrix}
\alias{applyAgeTimeMatrix}
\title{applies a age time matrix to a dataframe containing an incidence, an age category column, and a date column}
\usage{
applyAgeTimeMatrix(
  inputDf,
  matrixDf,
  oldIncidenceVar = "incidence",
  newIncidenceVar = "incidence",
  oldDateVar = "date",
  newDateVar = "date",
  ageCatVar = "ageGroup"
)
}
\arguments{
\item{inputDf}{- an input df containing an incidence col, a date col and a ageCategory col this must be grouped}

\item{matrixDf}{- a dataframe containing a sparse matrix of probabilities where age is labelled}

\item{oldIncidenceVar}{- the name of the input incidence col}

\item{newIncidenceVar}{- the name of the output incidence col}

\item{oldDateVar}{- the name of the input date column}

\item{newDateVar}{- the name of the output date column}

\item{ageCatVar}{- the age category variable}
}
\value{
a data frame containing the simulation results
}
\description{
Make sure your age categories are the saem as the matrix you are applying
}
\details{
The age time matrix can implement a delay, aggregate over time and or filter by age
}
