% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs-dataset-tools.R
\name{cleanseCHESSData}
\alias{cleanseCHESSData}
\title{Cleanses CHESS data set and adds a few fields}
\usage{
cleanseCHESSData(CHESSdf, date)
}
\arguments{
\item{path}{- path to the ff100 file}
}
\value{
cleansed CHESS data set
}
\description{
* Detects trusts that have not been updating records
* establishes a censoring date on a trust by trust basis
* checks all outcomes have an outcomedate
* checks outcome dates are after admission date
* excludes patients whose tests are >10 days after admission e.g. hospital acquired
* exludes paediatric patients, and those with no known gender
* adds censoring status (as status)
* adds time from admission to outcome (as timeToOutcome)
}
