% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-utils.R
\name{compute}
\alias{compute}
\title{Grouping preserving compute}
\usage{
compute(
  remote_df,
  name = NULL,
  temporary = FALSE,
  nocache = NULL,
  debug = NULL,
  params = NULL,
  prefix = NULL,
  ...
)
}
\arguments{
\item{remote_df}{}

\item{...}{}
}
\value{

}
\description{
The standard compute function messes with the grouping of a dataframe leading to weirdness. There are some 
other edge cases where grouping a data frame before an arrange causes issues. Some of these can be handled
by the window_order and associated functions
}
