% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-utils.R
\name{computeTemp}
\alias{computeTemp}
\title{Grouping preserving semi temporary compute}
\usage{
computeTemp(
  remote_df,
  prefix = "zz_tmp",
  temporary = FALSE,
  nocache = FALSE,
  params = NULL,
  ...
)
}
\arguments{
\item{remote_df}{}

\item{...}{}
}
\value{

}
\description{
The standard compute function messes with the grouping of a dataframe leading to weirdness. There are some 
other edge cases where grouping a data frame before an arrange causes issues. Some of these can be handled
by the window_order and associated functions
}
