% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIMDatasetProvider.R
\name{convertRtToSPIM}
\alias{convertRtToSPIM}
\title{reformats an Rt data set to SPI-M template}
\usage{
convertRtToSPIM(
  df,
  geographyExpr,
  modelExpr,
  modelTypeExpr,
  dateVar = "date",
  groupName = "Exeter",
  version = "0.1"
)
}
\arguments{
\item{df}{- the dataframe with Rt estimates}

\item{geographyExpr}{- the derivation of the geographic name}

\item{dateVar}{- the column with dates}

\item{groupName}{- the group submitting the estimate}

\item{modelName}{- the model}

\item{modelTypeExpr:}{ModelType - Multiple, Deaths, Cases, Survey, Emergency, Pillar Two Testing}
}
\value{
a tibble of the formatted estimates
}
\description{
reformats an Rt data set to SPI-M template
}
