% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovidTimeseriesProvider.R
\name{covidTimeseriesFormat}
\alias{covidTimeseriesFormat}
\title{Check timeseries conforms}
\usage{
covidTimeseriesFormat(.)
}
\value{
A dataframe that is specific to the covid timeseries format.
}
\description{
Check timeseries conforms
}
