% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptools.R
\name{createNeighbourNetwork}
\alias{createNeighbourNetwork}
\title{create a neighbourhood network from a shapefile}
\usage{
createNeighbourNetwork(map, idVar)
}
\arguments{
\item{map}{- a sf object defining regions / geometries - e.g. shapefile}

\item{idVar}{- the unique identifier for each region in the shapefile}
}
\value{
an edgelist of ids with from and to columns
}
\description{
create a neighbourhood network from a shapefile
}
