% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProvider.R
\name{DataProvider}
\alias{DataProvider}
\title{Caching and dataset management}
\description{
Caching and dataset management

Caching and dataset management
}
\section{Super class}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{DataProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataProvider$new()}}
\item \href{#method-downloadAndUnzip}{\code{DataProvider$downloadAndUnzip()}}
\item \href{#method-downloadAndUntar}{\code{DataProvider$downloadAndUntar()}}
\item \href{#method-download}{\code{DataProvider$download()}}
\item \href{#method-downloadDaily}{\code{DataProvider$downloadDaily()}}
\item \href{#method-normaliseGender}{\code{DataProvider$normaliseGender()}}
\item \href{#method-normaliseAgeCat}{\code{DataProvider$normaliseAgeCat()}}
\item \href{#method-cutByAge}{\code{DataProvider$cutByAge()}}
\item \href{#method-breakFromCats}{\code{DataProvider$breakFromCats()}}
\item \href{#method-labelsFromBreaks}{\code{DataProvider$labelsFromBreaks()}}
\item \href{#method-ageCatToFactor}{\code{DataProvider$ageCatToFactor()}}
\item \href{#method-ageToAgeCat}{\code{DataProvider$ageToAgeCat()}}
\item \href{#method-clone}{\code{DataProvider$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$new(providerController, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadAndUnzip"></a>}}
\subsection{Method \code{downloadAndUnzip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$downloadAndUnzip(id, url, pattern)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadAndUntar"></a>}}
\subsection{Method \code{downloadAndUntar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$downloadAndUntar(id, url, pattern)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\subsection{Method \code{download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$download(id, url, type = "csv")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadDaily"></a>}}
\subsection{Method \code{downloadDaily()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$downloadDaily(id, url, type = "csv")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normaliseGender"></a>}}
\subsection{Method \code{normaliseGender()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$normaliseGender(gender, na.value = NA_character_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normaliseAgeCat"></a>}}
\subsection{Method \code{normaliseAgeCat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$normaliseAgeCat(ageCat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cutByAge"></a>}}
\subsection{Method \code{cutByAge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$cutByAge(age, ageBreaks = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breakFromCats"></a>}}
\subsection{Method \code{breakFromCats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$breakFromCats(ageCat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-labelsFromBreaks"></a>}}
\subsection{Method \code{labelsFromBreaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$labelsFromBreaks(ageBreaks)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageCatToFactor"></a>}}
\subsection{Method \code{ageCatToFactor()}}{
ordered factor from age range labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$ageCatToFactor(
  ageCat,
  ageLabels = c("0-4", "<5", "5-9", "5-14", "10-14", "15-19", "15-24", "20-24", "25-29",
    "25-34", "30-34", "35-39", "35-44", "40-44", "45-49", "45-54", "50-54", "55-59",
    "55-64", "60-64", "65-69", "65-74", "70-74", "75-79", "75-84", "80+", "85+",
    "unknown")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ageCat}}{- a vector of age categories as strings}

\item{\code{ageLabels}}{- a vector of age range labels}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an ordered factor of age categories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageToAgeCat"></a>}}
\subsection{Method \code{ageToAgeCat()}}{
create an ordered factor of ages from a continuous age
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$ageToAgeCat(
  age,
  ageLabels = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{age}}{- a vector of ages}

\item{\code{ageLabels}}{- a vector of age range labels}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an ordered factor of age categories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
