% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs-dataset-tools.R
\name{generateSurvivalData}
\alias{generateSurvivalData}
\title{generrate survival data and right censored data from}
\usage{
generateSurvivalData(
  df,
  startDateVar,
  outcomeDateVar,
  outcomeExpr,
  ageVar = "age",
  ageBreaks = c(-Inf, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80,
    Inf),
  ageLabels = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+"),
  ageReferenceCat = NA
)
}
\arguments{
\item{path}{- path to the ff100 file}
}
\value{
data set in survival format including status, time, left, right, ageCat columns
}
\description{
generrate survival data and right censored data from
}
