% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs-dataset-tools.R
\name{getCHESS}
\alias{getCHESS}
\title{create a neighbourhood network from a shapefile}
\usage{
getCHESS(path)
}
\arguments{
\item{path}{- a path to the chess csv file}
}
\value{
raw CHESS data set
}
\description{
create a neighbourhood network from a shapefile
}
