% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs-dataset-tools.R
\name{getCHESSSummary}
\alias{getCHESSSummary}
\title{Load Chess summary file}
\usage{
getCHESSSummary(path)
}
\arguments{
\item{path}{- path to the ff100 file}
}
\value{
raw FF100 data set
}
\description{
Load Chess summary file
}
