% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs-dataset-tools.R
\name{getLineList}
\alias{getLineList}
\title{Load line list}
\usage{
getLineList(path)
}
\arguments{
\item{path}{- path to the line list file}
}
\value{
raw CHESS data set
}
\description{
Load line list
}
