% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcovidtools.R
\name{getUKCovidTimeseries}
\alias{getUKCovidTimeseries}
\title{Get UK outbreak timeseries data}
\usage{
getUKCovidTimeseries()
}
\value{
a data frame with several timeseries in it
}
\description{
data held in google sheets:
https://docs.google.com/spreadsheets/d/1snb-vYuH7fVpTuyoQrM8zWiABYoXbSrnn44w-zlhM90/edit?usp=sharing
}
