% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovidTimeseriesProvider.R
\name{\%def\%}
\alias{\%def\%}
\title{In fix operator to apply ensurer rules}
\usage{
ensurer \%def\% object
}
\arguments{
\item{ensurer}{an ensurer contract}

\item{object}{the object}
}
\value{
object if it conforms stops otherwise.
}
\description{
In fix operator to apply ensurer rules
}
