% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-utils.R
\name{\%like\%}
\alias{\%like\%}
\title{SQL like filtering in data tables}
\usage{
vector \%like\% like
}
\description{
allow a sql like syntax in data tables. N.b. conflict with data.table's definition of like which is
based on regex.
}
\details{
%	Represents zero or more characters	bl% finds bl, black, blue, and blob
_	Represents a single character	h_t finds hot, hat, and hit
[]	Represents any single character within the brackets	h[oa]t finds hot and hat, but not hit
^	Represents any character not in the brackets	h[^oa]t finds hit, but not hot and hat
-	Represents a range of characters	c[a-b]t finds cat and cbt
TODO: escaped sql
}
