% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcovidtools.R
\name{joinList}
\alias{joinList}
\title{Calculates a join list}
\usage{
joinList(df, groupVars = NULL, defaultJoin = NULL)
}
\arguments{
\item{df}{- a df which may be grouped}

\item{groupVars}{- the grouping for which we want to create a label as a list of columns quoted by vars(...)}

\item{defaultJoin}{- if there is no grouping we need one column to join by.}
}
\value{
a join List
}
\description{
Calculates a join list
}
