% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawardtools.R
\name{loadAgeTimeMatrix}
\alias{loadAgeTimeMatrix}
\title{loads an age time matrix from a csv}
\usage{
loadAgeTimeMatrix(url)
}
\arguments{
\item{url}{- a url or filename for a csv file}
}
\value{
a data frame containing the filter
}
\description{
loads an age time matrix from a csv
}
