% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawardtools.R
\name{loadDefaultAgeTimeMatrices}
\alias{loadDefaultAgeTimeMatrices}
\title{loads a ventilator demand matrices}
\usage{
loadDefaultAgeTimeMatrices()
}
\value{
a list of matrices
}
\description{
loads a ventilator demand matrices
}
