% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileProvider.R
\name{LocalFileProvider}
\alias{LocalFileProvider}
\title{Read only file providers}
\description{
Read only file providers

Read only file providers
}
\section{Super classes}{
\code{\link[ukcovidtools:AbstractFileProvider]{ukcovidtools::AbstractFileProvider}} -> \code{\link[ukcovidtools:ReadOnlyFileProvider]{ukcovidtools::ReadOnlyFileProvider}} -> \code{LocalFileProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-absolutePath}{\code{LocalFileProvider$absolutePath()}}
\item \href{#method-new}{\code{LocalFileProvider$new()}}
\item \href{#method-listFiles}{\code{LocalFileProvider$listFiles()}}
\item \href{#method-getFile}{\code{LocalFileProvider$getFile()}}
\item \href{#method-print}{\code{LocalFileProvider$print()}}
\item \href{#method-clone}{\code{LocalFileProvider$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="listAllFiles">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-listAllFiles}{\code{ukcovidtools::AbstractFileProvider$listAllFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="relativePath">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-relativePath}{\code{ukcovidtools::AbstractFileProvider$relativePath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="ReadOnlyFileProvider" data-id="processFiles">}\href{../../ukcovidtools/html/ReadOnlyFileProvider.html#method-processFiles}{\code{ukcovidtools::ReadOnlyFileProvider$processFiles()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-absolutePath"></a>}}
\subsection{Method \code{absolutePath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileProvider$absolutePath(parts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileProvider$new(config, root = config$root, sep = .Platform$file.sep)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listFiles"></a>}}
\subsection{Method \code{listFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileProvider$listFiles(directory = ".")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFile"></a>}}
\subsection{Method \code{getFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileProvider$getFile(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileProvider$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFileProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
