% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSDatasetProvider.R
\name{NHSDatasetProvider}
\alias{NHSDatasetProvider}
\title{NHS datasets}
\description{
NHS datasets

NHS datasets
}
\section{Super classes}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{\link[ukcovidtools:CovidTimeseriesProvider]{ukcovidtools::CovidTimeseriesProvider}} -> \code{NHSDatasetProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NHSDatasetProvider$new()}}
\item \href{#method-getGoogleMobility}{\code{NHSDatasetProvider$getGoogleMobility()}}
\item \href{#method-getSignificantDates}{\code{NHSDatasetProvider$getSignificantDates()}}
\item \href{#method-getPublicOneOneOne}{\code{NHSDatasetProvider$getPublicOneOneOne()}}
\item \href{#method-getTomWhiteCases}{\code{NHSDatasetProvider$getTomWhiteCases()}}
\item \href{#method-getTomWhiteIndicators}{\code{NHSDatasetProvider$getTomWhiteIndicators()}}
\item \href{#method-getNHSDeaths}{\code{NHSDatasetProvider$getNHSDeaths()}}
\item \href{#method-getPHEDashboard}{\code{NHSDatasetProvider$getPHEDashboard()}}
\item \href{#method-getTheFireHose}{\code{NHSDatasetProvider$getTheFireHose()}}
\item \href{#method-clone}{\code{NHSDatasetProvider$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="ageCatToFactor">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-ageCatToFactor}{\code{ukcovidtools::PassthroughFilesystemCache$ageCatToFactor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="ageToAgeCat">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-ageToAgeCat}{\code{ukcovidtools::PassthroughFilesystemCache$ageToAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="breakFromCats">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-breakFromCats}{\code{ukcovidtools::PassthroughFilesystemCache$breakFromCats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="cutByAge">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-cutByAge}{\code{ukcovidtools::PassthroughFilesystemCache$cutByAge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="download">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-download}{\code{ukcovidtools::PassthroughFilesystemCache$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="downloadAndUnzip">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-downloadAndUnzip}{\code{ukcovidtools::PassthroughFilesystemCache$downloadAndUnzip()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="downloadDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-downloadDaily}{\code{ukcovidtools::PassthroughFilesystemCache$downloadDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="importTimeseries">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-importTimeseries}{\code{ukcovidtools::PassthroughFilesystemCache$importTimeseries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="normaliseGender">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-normaliseGender}{\code{ukcovidtools::PassthroughFilesystemCache$normaliseGender()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="complete">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-complete}{\code{ukcovidtools::CovidTimeseriesProvider$complete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="describe">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-describe}{\code{ukcovidtools::CovidTimeseriesProvider$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="describeErrors">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-describeErrors}{\code{ukcovidtools::CovidTimeseriesProvider$describeErrors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="fixDatesAndNames">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-fixDatesAndNames}{\code{ukcovidtools::CovidTimeseriesProvider$fixDatesAndNames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="timeseriesQA">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-timeseriesQA}{\code{ukcovidtools::CovidTimeseriesProvider$timeseriesQA()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="trimNAs">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-trimNAs}{\code{ukcovidtools::CovidTimeseriesProvider$trimNAs()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$new(providerController, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGoogleMobility"></a>}}
\subsection{Method \code{getGoogleMobility()}}{
Load google mobility file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getGoogleMobility(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
google mobility 0 custom format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSignificantDates"></a>}}
\subsection{Method \code{getSignificantDates()}}{
Load curated significant dates file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getSignificantDates(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
public 111 data including pathways and online, pre and post april. Codes are CCG20 codes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPublicOneOneOne"></a>}}
\subsection{Method \code{getPublicOneOneOne()}}{
Load public 111 summary file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPublicOneOneOne(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
public 111 data including pathways and online, pre and post april. Codes are CCG20 codes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTomWhiteCases"></a>}}
\subsection{Method \code{getTomWhiteCases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getTomWhiteCases(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTomWhiteIndicators"></a>}}
\subsection{Method \code{getTomWhiteIndicators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getTomWhiteIndicators(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSDeaths"></a>}}
\subsection{Method \code{getNHSDeaths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getNHSDeaths(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPHEDashboard"></a>}}
\subsection{Method \code{getPHEDashboard()}}{
Get UK outbreak timeseries data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPHEDashboard(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeserisFormat data frame with several timeseries in it
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTheFireHose"></a>}}
\subsection{Method \code{getTheFireHose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getTheFireHose(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
