% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcovidtools.R
\name{normaliseAndCleanse}
\alias{normaliseAndCleanse}
\title{Cleanse UK outbreak timeseries data}
\usage{
normaliseAndCleanse(
  groupedDf,
  dateVar = "date",
  cumulativeCasesVar = "cumulative_cases",
  totalVar = "daily_total",
  unknownVar = "daily_unknown",
  adjustUnknowns = TRUE
)
}
\arguments{
\item{groupedDf}{the tidy dataframe grouped by regional geographical areas whatever you have}

\item{dateVar}{the column containing the date}

\item{cumulativeCasesVar}{}

\item{totalVar}{the total over all the whole geography}

\item{unknownVar}{}
}
\description{
Unofficial UK timeseries data has lots of data quality issues.
Various things done to fix:
* scales cumulative cases by unknown/total+1 across the board 
* calculates incidence and makes sure no negative incidences recorded
* updated cumulative_cases to reflect cleaned incidences
}
