% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassthroughFilesystemCache.R
\name{PassthroughFilesystemCache}
\alias{PassthroughFilesystemCache}
\title{Caching and dataset management}
\description{
Caching and dataset management

Caching and dataset management
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PassthroughFilesystemCache$new()}}
\item \href{#method-getDaily}{\code{PassthroughFilesystemCache$getDaily()}}
\item \href{#method-getHashCached}{\code{PassthroughFilesystemCache$getHashCached()}}
\item \href{#method-getSaved}{\code{PassthroughFilesystemCache$getSaved()}}
\item \href{#method-downloadAndUnzip}{\code{PassthroughFilesystemCache$downloadAndUnzip()}}
\item \href{#method-download}{\code{PassthroughFilesystemCache$download()}}
\item \href{#method-downloadDaily}{\code{PassthroughFilesystemCache$downloadDaily()}}
\item \href{#method-normaliseGender}{\code{PassthroughFilesystemCache$normaliseGender()}}
\item \href{#method-cutByAge}{\code{PassthroughFilesystemCache$cutByAge()}}
\item \href{#method-breakFromCats}{\code{PassthroughFilesystemCache$breakFromCats()}}
\item \href{#method-ageCatToFactor}{\code{PassthroughFilesystemCache$ageCatToFactor()}}
\item \href{#method-ageToAgeCat}{\code{PassthroughFilesystemCache$ageToAgeCat()}}
\item \href{#method-importTimeseries}{\code{PassthroughFilesystemCache$importTimeseries()}}
\item \href{#method-clone}{\code{PassthroughFilesystemCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$new(
  providerController,
  nocache = FALSE,
  debug = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDaily"></a>}}
\subsection{Method \code{getDaily()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$getDaily(id, orElse, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHashCached"></a>}}
\subsection{Method \code{getHashCached()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$getHashCached(
  object,
  operation,
  params = NULL,
  orElse,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSaved"></a>}}
\subsection{Method \code{getSaved()}}{
a pass through 2 level cache (memory / saved file / orElse function)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$getSaved(id, orElse, params = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadAndUnzip"></a>}}
\subsection{Method \code{downloadAndUnzip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$downloadAndUnzip(id, url, pattern)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\subsection{Method \code{download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$download(id, url, type = "csv")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadDaily"></a>}}
\subsection{Method \code{downloadDaily()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$downloadDaily(id, url, type = "csv")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normaliseGender"></a>}}
\subsection{Method \code{normaliseGender()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$normaliseGender(gender)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cutByAge"></a>}}
\subsection{Method \code{cutByAge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$cutByAge(age, ageBreaks = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breakFromCats"></a>}}
\subsection{Method \code{breakFromCats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$breakFromCats(ageCat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageCatToFactor"></a>}}
\subsection{Method \code{ageCatToFactor()}}{
ordered factor from age range labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$ageCatToFactor(
  ageCat,
  ageLabels = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ageCat}}{- a vector of age categories as strings}

\item{\code{ageLabels}}{- a vector of age range labels}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an ordered factor of age categories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageToAgeCat"></a>}}
\subsection{Method \code{ageToAgeCat()}}{
create an ordered factor of ages from a continuous age
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$ageToAgeCat(
  age,
  ageLabels = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{age}}{- a vector of ages}

\item{\code{ageLabels}}{- a vector of age range labels}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an ordered factor of age categories
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importTimeseries"></a>}}
\subsection{Method \code{importTimeseries()}}{
Calculate an estimate of rate of change of Rt using a loess
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$importTimeseries(
  df,
  dateVar,
  valueExpr,
  statisticExpr,
  typeExpr,
  codeExpr = "no code",
  nameExpr = "no name",
  codeTypeExpr = "undefined",
  sourceExpr = "undefined",
  ageCatExpr = NA,
  genderExpr = NA,
  subgroupExpr = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{R0timeseries}}{a grouped df containing R0 timeseries including a date and a Median(R) column from EpiEstim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PassthroughFilesystemCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
