% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-utils.R
\name{plotRibbons}
\alias{plotRibbons}
\title{Add ribbons to times series from mean and SD}
\usage{
plotRibbons(meanVar, sdVar, colourExpr, ...)
}
\arguments{
\item{meanVar}{the mean}

\item{sdVar}{the sd}
}
\description{
Add ribbons to times series from mean and SD
}
