% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcovidtools.R
\name{print_and_capture}
\alias{print_and_capture}
\title{capture a data fram for an error message}
\usage{
print_and_capture(x)
}
\arguments{
\item{x}{a dataframe}
}
\description{
capture a data fram for an error message
}
