% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-utils.R
\name{printDataframeToString}
\alias{printDataframeToString}
\title{#' colApply apply the function to the rows of a dataframe row-wise
#' 
#' @param f - a function which operates on a vector
#' @param ... a list of vectors
colApply = function(f, ...) {
  browser()
  if (!all(sapply(list(...), length)==length(list(...)[[1]]))) stop("arguments must all have same length")
  m = matrix(c(...),ncol=length(list(...)))
  out = apply(m, 1, f, na.rm=TRUE)
  return(out)
}
capture a data frame for an error message}
\usage{
printDataframeToString(x)
}
\arguments{
\item{x}{a dataframe}
}
\description{
#' colApply apply the function to the rows of a dataframe row-wise
#' 
#' @param f - a function which operates on a vector
#' @param ... a list of vectors
colApply = function(f, ...) {
  browser()
  if (!all(sapply(list(...), length)==length(list(...)[[1]]))) stop("arguments must all have same length")
  m = matrix(c(...),ncol=length(list(...)))
  out = apply(m, 1, f, na.rm=TRUE)
  return(out)
}
capture a data frame for an error message
}
