% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawardtools.R
\name{readSpatialFile}
\alias{readSpatialFile}
\title{loads a spatial model file}
\usage{
readSpatialFile(fname, ignoreErrors = FALSE)
}
\arguments{
\item{fname}{- a filename containing a ward based metaward model .dat file}
}
\value{
a data frame containing the simulation results
}
\description{
loads a spatial model file
}
