% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileProvider.R
\name{S3FileProvider}
\alias{S3FileProvider}
\title{Read only file providers}
\description{
Read only file providers

Read only file providers
}
\section{Super classes}{
\code{\link[ukcovidtools:AbstractFileProvider]{ukcovidtools::AbstractFileProvider}} -> \code{\link[ukcovidtools:ReadOnlyFileProvider]{ukcovidtools::ReadOnlyFileProvider}} -> \code{S3FileProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{S3FileProvider$new()}}
\item \href{#method-finalize}{\code{S3FileProvider$finalize()}}
\item \href{#method-listAllFiles}{\code{S3FileProvider$listAllFiles()}}
\item \href{#method-listFiles}{\code{S3FileProvider$listFiles()}}
\item \href{#method-getFile}{\code{S3FileProvider$getFile()}}
\item \href{#method-print}{\code{S3FileProvider$print()}}
\item \href{#method-clone}{\code{S3FileProvider$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="relativePath">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-relativePath}{\code{ukcovidtools::AbstractFileProvider$relativePath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="ReadOnlyFileProvider" data-id="processFiles">}\href{../../ukcovidtools/html/ReadOnlyFileProvider.html#method-processFiles}{\code{ukcovidtools::ReadOnlyFileProvider$processFiles()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
SFTP over SSH tunnel file provider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3FileProvider$new(
  config,
  bucket = config$bucket,
  key = config$accesskey,
  secret = config$secretkey,
  region = config$region,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{- config object}

\item{\code{bucket}}{- S3 bucket name}

\item{\code{key}}{- S3 access key}

\item{\code{secret}}{- S3 secret key}

\item{\code{region}}{- S3 region}

\item{\code{...}}{for compatibility}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the provider
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3FileProvider$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listAllFiles"></a>}}
\subsection{Method \code{listAllFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3FileProvider$listAllFiles(directory = ".")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listFiles"></a>}}
\subsection{Method \code{listFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3FileProvider$listFiles(directory = ".")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFile"></a>}}
\subsection{Method \code{getFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3FileProvider$getFile(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3FileProvider$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3FileProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
