% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-utils.R
\name{sdFromProduct}
\alias{sdFromProduct}
\title{Calculates a standard deviation for a product distribution of gaussians}
\usage{
sdFromProduct(mu_x, sig_x, mu_y = mu_x, sig_y = sig_x)
}
\value{
a SD
}
\description{
Calculates a standard deviation for a product distribution of gaussians
}
