% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-utils.R
\name{sdFromRatio}
\alias{sdFromRatio}
\title{Calculates a standard deviation for a ratio distribution of gaussians}
\usage{
sdFromRatio(mu_x, sig_x, mu_y, sig_y)
}
\value{
a SD
}
\description{
Calculates a standard deviation for a ratio distribution of gaussians
}
