% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileProvider.R
\name{SFTPFileProvider}
\alias{SFTPFileProvider}
\title{Read only file providers}
\description{
Read only file providers

Read only file providers
}
\section{Super classes}{
\code{\link[ukcovidtools:AbstractFileProvider]{ukcovidtools::AbstractFileProvider}} -> \code{\link[ukcovidtools:ReadOnlyFileProvider]{ukcovidtools::ReadOnlyFileProvider}} -> \code{SFTPFileProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SFTPFileProvider$new()}}
\item \href{#method-print}{\code{SFTPFileProvider$print()}}
\item \href{#method-clone}{\code{SFTPFileProvider$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="listAllFiles">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-listAllFiles}{\code{ukcovidtools::AbstractFileProvider$listAllFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="listFiles">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-listFiles}{\code{ukcovidtools::AbstractFileProvider$listFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="relativePath">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-relativePath}{\code{ukcovidtools::AbstractFileProvider$relativePath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="ReadOnlyFileProvider" data-id="getFile">}\href{../../ukcovidtools/html/ReadOnlyFileProvider.html#method-getFile}{\code{ukcovidtools::ReadOnlyFileProvider$getFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="ReadOnlyFileProvider" data-id="processFiles">}\href{../../ukcovidtools/html/ReadOnlyFileProvider.html#method-processFiles}{\code{ukcovidtools::ReadOnlyFileProvider$processFiles()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPFileProvider$new(
  config,
  hostName = config$hostName,
  user = config$user,
  password = config$password
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPFileProvider$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPFileProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
