% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileProvider.R
\name{SFTPOverSSHFileProvider}
\alias{SFTPOverSSHFileProvider}
\title{Read only file providers}
\description{
Read only file providers

Read only file providers
}
\section{Super classes}{
\code{\link[ukcovidtools:AbstractFileProvider]{ukcovidtools::AbstractFileProvider}} -> \code{\link[ukcovidtools:ReadOnlyFileProvider]{ukcovidtools::ReadOnlyFileProvider}} -> \code{SFTPOverSSHFileProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SFTPOverSSHFileProvider$new()}}
\item \href{#method-finalize}{\code{SFTPOverSSHFileProvider$finalize()}}
\item \href{#method-curlCommand}{\code{SFTPOverSSHFileProvider$curlCommand()}}
\item \href{#method-getSession}{\code{SFTPOverSSHFileProvider$getSession()}}
\item \href{#method-listFiles}{\code{SFTPOverSSHFileProvider$listFiles()}}
\item \href{#method-getFile}{\code{SFTPOverSSHFileProvider$getFile()}}
\item \href{#method-print}{\code{SFTPOverSSHFileProvider$print()}}
\item \href{#method-clone}{\code{SFTPOverSSHFileProvider$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="listAllFiles">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-listAllFiles}{\code{ukcovidtools::AbstractFileProvider$listAllFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="relativePath">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-relativePath}{\code{ukcovidtools::AbstractFileProvider$relativePath()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="ReadOnlyFileProvider" data-id="processFiles">}\href{../../ukcovidtools/html/ReadOnlyFileProvider.html#method-processFiles}{\code{ukcovidtools::ReadOnlyFileProvider$processFiles()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
SFTP over SSH tunnel file provider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPOverSSHFileProvider$new(
  config,
  hostName = config$hostName,
  ssh = config$ssh,
  sshkey = config$sshkey,
  user = config$user,
  password = config$password,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ssh}}{- ssh tunnel host}

\item{\code{sshkey}}{- ssh key file}

\item{\code{user}}{- sftp username}

\item{\code{password}}{- sftp password}

\item{\code{...}}{for compatibility}

\item{\code{hostname}}{- sftp host}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the provider
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPOverSSHFileProvider$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-curlCommand"></a>}}
\subsection{Method \code{curlCommand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPOverSSHFileProvider$curlCommand(relativePath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSession"></a>}}
\subsection{Method \code{getSession()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPOverSSHFileProvider$getSession()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listFiles"></a>}}
\subsection{Method \code{listFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPOverSSHFileProvider$listFiles(directory = ".")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFile"></a>}}
\subsection{Method \code{getFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPOverSSHFileProvider$getFile(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPOverSSHFileProvider$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFTPOverSSHFileProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
