% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIMDatasetProvider.R
\name{SPIMDatasetProvider}
\alias{SPIMDatasetProvider}
\title{SPIM private data}
\description{
SPIM private data

SPIM private data
}
\section{Super classes}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{\link[ukcovidtools:CovidTimeseriesProvider]{ukcovidtools::CovidTimeseriesProvider}} -> \code{SPIMDatasetProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SPIMDatasetProvider$new()}}
\item \href{#method-getPaths}{\code{SPIMDatasetProvider$getPaths()}}
\item \href{#method-getDeathsLineList}{\code{SPIMDatasetProvider$getDeathsLineList()}}
\item \href{#method-getLineList}{\code{SPIMDatasetProvider$getLineList()}}
\item \href{#method-getSeroprevalence}{\code{SPIMDatasetProvider$getSeroprevalence()}}
\item \href{#method-getFF100}{\code{SPIMDatasetProvider$getFF100()}}
\item \href{#method-getBristolData}{\code{SPIMDatasetProvider$getBristolData()}}
\item \href{#method-getCHESS}{\code{SPIMDatasetProvider$getCHESS()}}
\item \href{#method-getCHESSSummary}{\code{SPIMDatasetProvider$getCHESSSummary()}}
\item \href{#method-bristolSurvivalSubset}{\code{SPIMDatasetProvider$bristolSurvivalSubset()}}
\item \href{#method-getLineListIncidence}{\code{SPIMDatasetProvider$getLineListIncidence()}}
\item \href{#method-getDeathsLineListIncidence}{\code{SPIMDatasetProvider$getDeathsLineListIncidence()}}
\item \href{#method-getSeroprevalenceTestIncidence}{\code{SPIMDatasetProvider$getSeroprevalenceTestIncidence()}}
\item \href{#method-getOneOneOne}{\code{SPIMDatasetProvider$getOneOneOne()}}
\item \href{#method-getSPIMextract}{\code{SPIMDatasetProvider$getSPIMextract()}}
\item \href{#method-getTheSPIMFireHose}{\code{SPIMDatasetProvider$getTheSPIMFireHose()}}
\item \href{#method-clone}{\code{SPIMDatasetProvider$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="ageCatToFactor">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-ageCatToFactor}{\code{ukcovidtools::PassthroughFilesystemCache$ageCatToFactor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="ageToAgeCat">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-ageToAgeCat}{\code{ukcovidtools::PassthroughFilesystemCache$ageToAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="breakFromCats">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-breakFromCats}{\code{ukcovidtools::PassthroughFilesystemCache$breakFromCats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="cutByAge">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-cutByAge}{\code{ukcovidtools::PassthroughFilesystemCache$cutByAge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="download">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-download}{\code{ukcovidtools::PassthroughFilesystemCache$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="downloadAndUnzip">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-downloadAndUnzip}{\code{ukcovidtools::PassthroughFilesystemCache$downloadAndUnzip()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="downloadDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-downloadDaily}{\code{ukcovidtools::PassthroughFilesystemCache$downloadDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="importTimeseries">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-importTimeseries}{\code{ukcovidtools::PassthroughFilesystemCache$importTimeseries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="normaliseGender">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-normaliseGender}{\code{ukcovidtools::PassthroughFilesystemCache$normaliseGender()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="complete">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-complete}{\code{ukcovidtools::CovidTimeseriesProvider$complete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="describe">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-describe}{\code{ukcovidtools::CovidTimeseriesProvider$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="describeErrors">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-describeErrors}{\code{ukcovidtools::CovidTimeseriesProvider$describeErrors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="fixDatesAndNames">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-fixDatesAndNames}{\code{ukcovidtools::CovidTimeseriesProvider$fixDatesAndNames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="timeseriesQA">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-timeseriesQA}{\code{ukcovidtools::CovidTimeseriesProvider$timeseriesQA()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="trimNAs">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-trimNAs}{\code{ukcovidtools::CovidTimeseriesProvider$trimNAs()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$new(providerController, path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPaths"></a>}}
\subsection{Method \code{getPaths()}}{
Search a file path for the
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getPaths(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the line list file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
raw line list data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDeathsLineList"></a>}}
\subsection{Method \code{getDeathsLineList()}}{
Load line list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getDeathsLineList(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the line list file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
raw line list data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLineList"></a>}}
\subsection{Method \code{getLineList()}}{
Load line list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getLineList(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw line list data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSeroprevalence"></a>}}
\subsection{Method \code{getSeroprevalence()}}{
Load the seroprevalance file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getSeroprevalence(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw FF100 data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFF100"></a>}}
\subsection{Method \code{getFF100()}}{
Load ff100 file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getFF100()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raw FF100 data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBristolData"></a>}}
\subsection{Method \code{getBristolData()}}{
Load Bristol data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getBristolData(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the bristol data file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
raw Bristol data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCHESS"></a>}}
\subsection{Method \code{getCHESS()}}{
Load the CHESS dataset from a path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getCHESS()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- a path to the chess csv file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
raw CHESS data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCHESSSummary"></a>}}
\subsection{Method \code{getCHESSSummary()}}{
Load Chess summary file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getCHESSSummary()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the ff100 file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
raw FF100 data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bristolSurvivalSubset"></a>}}
\subsection{Method \code{bristolSurvivalSubset()}}{
Load Bristol survival subset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$bristolSurvivalSubset(bristolDf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the ff100 file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Bristol survival subset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLineListIncidence"></a>}}
\subsection{Method \code{getLineListIncidence()}}{
Load incidence from line list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getLineListIncidence(
  ageBreaks = NULL,
  specimenOrReport = "specimen",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ageBreaks}}{- a list of ages which form the cut points for breaking continuous ages into ranges (or NULL for a single age category)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDeathsLineListIncidence"></a>}}
\subsection{Method \code{getDeathsLineListIncidence()}}{
Load deaths data from linelist - does not preserve ethnicity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getDeathsLineListIncidence(
  ageBreaks = NULL,
  deathOrReport = "death",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ageBreaks}}{- a list of ages which form the cut points for breaking continuous ages into ranges (or NULL for a single age category)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSeroprevalenceTestIncidence"></a>}}
\subsection{Method \code{getSeroprevalenceTestIncidence()}}{
Load seroprevalence data from linelist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getSeroprevalenceTestIncidence(ageBreaks = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ageBreaks}}{- a list of ages which form the cut points for breaking continuous ages into ranges (or NULL for a single age category)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOneOneOne"></a>}}
\subsection{Method \code{getOneOneOne()}}{
Load 111 data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getOneOneOne(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSPIMextract"></a>}}
\subsection{Method \code{getSPIMextract()}}{
Load the SPI-M aggregated data spreadsheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getSPIMextract(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTheSPIMFireHose"></a>}}
\subsection{Method \code{getTheSPIMFireHose()}}{
Load the SPI-M and public data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$getTheSPIMFireHose(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPIMDatasetProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
