% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalProcessingPipeline.R
\name{SurvivalProcessingPipeline}
\alias{SurvivalProcessingPipeline}
\title{Survival data procesing}
\description{
Survival data procesing

Survival data procesing
}
\section{Super class}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{SurvivalProcessingPipeline}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurvivalProcessingPipeline$new()}}
\item \href{#method-generateSurvivalData}{\code{SurvivalProcessingPipeline$generateSurvivalData()}}
\item \href{#method-generateNoAgeSurvivalData}{\code{SurvivalProcessingPipeline$generateNoAgeSurvivalData()}}
\item \href{#method-getMaxDate}{\code{SurvivalProcessingPipeline$getMaxDate()}}
\item \href{#method-transitionMatrix}{\code{SurvivalProcessingPipeline$transitionMatrix()}}
\item \href{#method-generateMultistateSurvivalData}{\code{SurvivalProcessingPipeline$generateMultistateSurvivalData()}}
\item \href{#method-createSurvivalSurfaces}{\code{SurvivalProcessingPipeline$createSurvivalSurfaces()}}
\item \href{#method-plotSurvivalSurface}{\code{SurvivalProcessingPipeline$plotSurvivalSurface()}}
\item \href{#method-plotIncidenceSurface}{\code{SurvivalProcessingPipeline$plotIncidenceSurface()}}
\item \href{#method-plotProbabilityMatrix}{\code{SurvivalProcessingPipeline$plotProbabilityMatrix()}}
\item \href{#method-clone}{\code{SurvivalProcessingPipeline$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$new(providerController, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateSurvivalData"></a>}}
\subsection{Method \code{generateSurvivalData()}}{
generate survival data and right censored data from
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$generateSurvivalData(
  df,
  idVar,
  startDateVar,
  endDateExpr,
  statusExpr = if_else(is.na({     {         endDateExpr     } }), 0, 1),
  censoredDateExpr = max({     {         endDateExpr     } }, na.rm = TRUE),
  ageVar = "age",
  ageBreaks = c(-Inf, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80,
    Inf),
  ageLabels = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+"),
  ageReferenceCat = NA_character_,
  statusLabels = c("censored", "died")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the ff100 file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data set in survival format including status, time, left, right, ageCat columns
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateNoAgeSurvivalData"></a>}}
\subsection{Method \code{generateNoAgeSurvivalData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$generateNoAgeSurvivalData(
  df,
  idVar,
  startDateVar,
  endDateExpr,
  statusExpr = if_else(is.na({     {         endDateExpr     } }), 0, 1),
  censoredDateExpr = max({     {         endDateExpr     } }, na.rm = TRUE),
  statusLabels = c("censored", "died")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMaxDate"></a>}}
\subsection{Method \code{getMaxDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$getMaxDate(df, dateCols)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transitionMatrix"></a>}}
\subsection{Method \code{transitionMatrix()}}{
generate transition matrix with enquoted tidy column sepecifications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$transitionMatrix(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{a set of mappings}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a transistion matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateMultistateSurvivalData"></a>}}
\subsection{Method \code{generateMultistateSurvivalData()}}{
generate survival data and right censored data from a multistate model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$generateMultistateSurvivalData(
  df,
  idVar,
  transMatrix,
  ageVar = "age",
  censoringDateExpr = getMaxDate(df, dateCols),
  startDateCol = names(allowableTransitions)[1],
  ageBreaks = c(-Inf, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80,
    Inf),
  ageLabels = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+"),
  ageReferenceCat = NA_character_,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the ff100 file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data set in survival format including status, time, left, right, ageCat columns
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createSurvivalSurfaces"></a>}}
\subsection{Method \code{createSurvivalSurfaces()}}{
for a set of model parameters, calculate the distributions and median
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$createSurvivalSurfaces(
  modelParameterDf,
  days = 30,
  timepoints = seq(0.1, days, length.out = 100)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelParameterDf}}{- a data frame set of parameters for a range of models as param, value columns}

\item{\code{days}}{- the time frame of the prediction}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an dataframe of ageCat and time varying surfaces, including pdfs, cdfs and medians, with high and low estimates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSurvivalSurface"></a>}}
\subsection{Method \code{plotSurvivalSurface()}}{
plot a survival surface as a CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$plotSurvivalSurface(surfacesDf, distr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surfacesDf}}{- an dataframe of ageCat and time varying surfaces, including pdfs, cdfs and medians, with high and low estimates}

\item{\code{distr}}{- a single distribution to plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a plot of the survival surface (1-cdf)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotIncidenceSurface"></a>}}
\subsection{Method \code{plotIncidenceSurface()}}{
plot a survival surface as a PDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$plotIncidenceSurface(surfacesDf, distr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surfacesDf}}{- an dataframe of ageCat and time varying surfaces, including pdfs, cdfs and medians, with high and low estimates}

\item{\code{distr}}{- a single distribution to plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a plot of the survival surface as pdf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotProbabilityMatrix"></a>}}
\subsection{Method \code{plotProbabilityMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$plotProbabilityMatrix(
  surfacesDf,
  distr,
  pExpr = "p"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
