% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcovidtools.R
\name{tidyEstimateRt}
\alias{tidyEstimateRt}
\title{Calculates a survival R(t) curve on grouped data}
\usage{
tidyEstimateRt(
  groupedDf,
  config,
  dateVar = "date",
  incidenceVar = "incidence",
  window = 2,
  ...
)
}
\arguments{
\item{groupedDf}{- a df which may be grouped, of timeseries data including an incidence column with continuous positive non NA values.}

\item{config}{An object of class estimate_R_config, as returned by function EpiEstim::make_config.}

\item{dateVar}{- the variable containing the seqence of dates}

\item{incidenceVar}{- the sequence of daily incidence}

\item{window}{- the width of the smoothing function applied (default 2)}
}
\value{
a dataframe with groupwise Rt estimates
}
\description{
N.B. EpiEstim has a bug affecting 0.25 quantiles which is being looked at
}
