% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukcovidtools-PACKAGE.R
\docType{package}
\name{ukcovidtools-package}
\alias{ukcovidtools}
\alias{ukcovidtools-package}
\title{ukcovidtools: Functions for sourcing, cleaning and calculating Rt for uk covid timeseries data}
\description{
What the package does (one paragraph).
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()]. 
Otherwise refer to the vignettes to see how to format the documentation.
}
\author{
\strong{Maintainer}: Robert Challen \email{rc538@exeter.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
