% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChessProcessingPipeline.R
\name{ChessProcessingPipeline}
\alias{ChessProcessingPipeline}
\title{Process chess data}
\description{
Process chess data

Process chess data
}
\section{Super classes}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{\link[ukcovidtools:DataProvider]{ukcovidtools::DataProvider}} -> \code{ChessProcessingPipeline}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ChessProcessingPipeline$new()}}
\item \href{#method-chessQuality}{\code{ChessProcessingPipeline$chessQuality()}}
\item \href{#method-chessItuSubset}{\code{ChessProcessingPipeline$chessItuSubset()}}
\item \href{#method-chessAdmissionSubset}{\code{ChessProcessingPipeline$chessAdmissionSubset()}}
\item \href{#method-chessDefaultFilter}{\code{ChessProcessingPipeline$chessDefaultFilter()}}
\item \href{#method-cleanseCHESSData}{\code{ChessProcessingPipeline$cleanseCHESSData()}}
\item \href{#method-clone}{\code{ChessProcessingPipeline$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="ageCatToFactor">}\href{../../ukcovidtools/html/DataProvider.html#method-ageCatToFactor}{\code{ukcovidtools::DataProvider$ageCatToFactor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="ageToAgeCat">}\href{../../ukcovidtools/html/DataProvider.html#method-ageToAgeCat}{\code{ukcovidtools::DataProvider$ageToAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="breakFromCats">}\href{../../ukcovidtools/html/DataProvider.html#method-breakFromCats}{\code{ukcovidtools::DataProvider$breakFromCats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="cutByAge">}\href{../../ukcovidtools/html/DataProvider.html#method-cutByAge}{\code{ukcovidtools::DataProvider$cutByAge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="download">}\href{../../ukcovidtools/html/DataProvider.html#method-download}{\code{ukcovidtools::DataProvider$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadAndUntar">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadAndUntar}{\code{ukcovidtools::DataProvider$downloadAndUntar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadAndUnzip">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadAndUnzip}{\code{ukcovidtools::DataProvider$downloadAndUnzip()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadDaily">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadDaily}{\code{ukcovidtools::DataProvider$downloadDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="labelsFromBreaks">}\href{../../ukcovidtools/html/DataProvider.html#method-labelsFromBreaks}{\code{ukcovidtools::DataProvider$labelsFromBreaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseAgeCat">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseAgeCat}{\code{ukcovidtools::DataProvider$normaliseAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseGender">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseGender}{\code{ukcovidtools::DataProvider$normaliseGender()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChessProcessingPipeline$new(providerController, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chessQuality"></a>}}
\subsection{Method \code{chessQuality()}}{
Chess data metrics on a hospital by hospital basis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChessProcessingPipeline$chessQuality(
  CHESSdf,
  updatedWithin = 14,
  date = max(CHESSdf$dateupdated, na.rm = TRUE),
  unknownDatesPercent = 0.2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CHESSdf}}{- the raw chess data set}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
the chess data has some issues as
different hospitals have different reporting strategies
}

\subsection{Returns}{
a data frame of stats by trust representing reporting quality.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chessItuSubset"></a>}}
\subsection{Method \code{chessItuSubset()}}{
The subset of the CHESS data of patients which are admitted to ITU

the chess data has some issues as
different hospitals have different reporting strategies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChessProcessingPipeline$chessItuSubset(
  CHESSdf = self$getCHESS(),
  updatedWithin = 14,
  date = max(CHESSdf$dateupdated, na.rm = TRUE),
  unknownDatesPercent = 0.2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CHESSdf}}{- the raw chess data set}

\item{\code{date}}{- the date of the data set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame of stats by trust representing reporting quality.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chessAdmissionSubset"></a>}}
\subsection{Method \code{chessAdmissionSubset()}}{
The subset of the CHESS data of patients from hospitals that appear to be reporting all admissions

the chess data has some issues as
different hospitals have different reporting strategies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChessProcessingPipeline$chessAdmissionSubset(
  CHESSdf = self$getCHESS(),
  updatedWithin = 14,
  date = max(CHESSdf$dateupdated, na.rm = TRUE),
  unknownDatesPercent = 0.2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CHESSdf}}{- the raw chess data set}

\item{\code{date}}{- the date of the data set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame of stats by trust representing reporting quality.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chessDefaultFilter"></a>}}
\subsection{Method \code{chessDefaultFilter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChessProcessingPipeline$chessDefaultFilter(CHESSdf = self$getCHESS(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanseCHESSData"></a>}}
\subsection{Method \code{cleanseCHESSData()}}{
Cleanses CHESS data set and adds a few fields

* Detects trusts that have not been updating records
* establishes a censoring date on a trust by trust basis
* checks all outcomes have an outcomedate
* checks outcome dates are after admission date
* excludes patients whose tests are >10 days after admission e.g. hospital acquired
* excludes paediatric patients, and those with no known gender
* adds censoring status (as status)
* adds time from admission to outcome (as timeToOutcome)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChessProcessingPipeline$cleanseCHESSData(CHESSdf, date, removeOutliers = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the ff100 file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cleansed CHESS data set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChessProcessingPipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
