% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIMDatasetProvider.R
\name{convertSerialIntervalToSPIM}
\alias{convertSerialIntervalToSPIM}
\title{reformats an EpiEstim config file to SPI-M template}
\usage{
convertSerialIntervalToSPIM(
  serial,
  modelName,
  groupName = "Exeter",
  version = "0.1"
)
}
\arguments{
\item{modelName}{- the model}

\item{groupName}{- the group submitting the estimate}

\item{cfg}{- the serial interval object}
}
\value{
a tibble of the formatted estimates
}
\description{
reformats an EpiEstim config file to SPI-M template
}
