% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-types.R
\name{covidTimeseriesFormat}
\alias{covidTimeseriesFormat}
\title{The covid timeseries format}
\usage{
covidTimeseriesFormat(.)
}
\value{
A dataframe that is specific to the covid timeseries format.
}
\description{
The covid timeseries format
}
