% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovidTimeseriesProvider.R
\name{CovidTimeseriesProvider}
\alias{CovidTimeseriesProvider}
\title{General covid  processing}
\description{
General covid  processing

General covid  processing
}
\section{Super classes}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{\link[ukcovidtools:DataProvider]{ukcovidtools::DataProvider}} -> \code{CovidTimeseriesProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CovidTimeseriesProvider$new()}}
\item \href{#method-setTruncation}{\code{CovidTimeseriesProvider$setTruncation()}}
\item \href{#method-trimNAs}{\code{CovidTimeseriesProvider$trimNAs()}}
\item \href{#method-complete}{\code{CovidTimeseriesProvider$complete()}}
\item \href{#method-completeAllRegions}{\code{CovidTimeseriesProvider$completeAllRegions()}}
\item \href{#method-fillAbsent}{\code{CovidTimeseriesProvider$fillAbsent()}}
\item \href{#method-fillAbsentAllRegions}{\code{CovidTimeseriesProvider$fillAbsentAllRegions()}}
\item \href{#method-fillAbsentByRegion}{\code{CovidTimeseriesProvider$fillAbsentByRegion()}}
\item \href{#method-fixDatesAndNames}{\code{CovidTimeseriesProvider$fixDatesAndNames()}}
\item \href{#method-fixDates}{\code{CovidTimeseriesProvider$fixDates()}}
\item \href{#method-timeseriesQA}{\code{CovidTimeseriesProvider$timeseriesQA()}}
\item \href{#method-describeErrors}{\code{CovidTimeseriesProvider$describeErrors()}}
\item \href{#method-describe}{\code{CovidTimeseriesProvider$describe()}}
\item \href{#method-clone}{\code{CovidTimeseriesProvider$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="unloadCache">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-unloadCache}{\code{ukcovidtools::PassthroughFilesystemCache$unloadCache()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="ageCatToFactor">}\href{../../ukcovidtools/html/DataProvider.html#method-ageCatToFactor}{\code{ukcovidtools::DataProvider$ageCatToFactor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="ageToAgeCat">}\href{../../ukcovidtools/html/DataProvider.html#method-ageToAgeCat}{\code{ukcovidtools::DataProvider$ageToAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="breakFromCats">}\href{../../ukcovidtools/html/DataProvider.html#method-breakFromCats}{\code{ukcovidtools::DataProvider$breakFromCats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="cutByAge">}\href{../../ukcovidtools/html/DataProvider.html#method-cutByAge}{\code{ukcovidtools::DataProvider$cutByAge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="download">}\href{../../ukcovidtools/html/DataProvider.html#method-download}{\code{ukcovidtools::DataProvider$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadAndUntar">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadAndUntar}{\code{ukcovidtools::DataProvider$downloadAndUntar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadAndUnzip">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadAndUnzip}{\code{ukcovidtools::DataProvider$downloadAndUnzip()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadDaily">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadDaily}{\code{ukcovidtools::DataProvider$downloadDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="labelsFromBreaks">}\href{../../ukcovidtools/html/DataProvider.html#method-labelsFromBreaks}{\code{ukcovidtools::DataProvider$labelsFromBreaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseAgeCat">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseAgeCat}{\code{ukcovidtools::DataProvider$normaliseAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseDemographics">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseDemographics}{\code{ukcovidtools::DataProvider$normaliseDemographics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseGender">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseGender}{\code{ukcovidtools::DataProvider$normaliseGender()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
New timeseries provider pipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$new(providerController, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{providerController}}{the provider controller}

\item{\code{...}}{for compatibility}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the provider
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTruncation"></a>}}
\if{latex}{\out{\hypertarget{method-setTruncation}{}}}
\subsection{Method \code{setTruncation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$setTruncation(listDays)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trimNAs"></a>}}
\if{latex}{\out{\hypertarget{method-trimNAs}{}}}
\subsection{Method \code{trimNAs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$trimNAs(r0Timeseries)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-complete"></a>}}
\if{latex}{\out{\hypertarget{method-complete}{}}}
\subsection{Method \code{complete()}}{
Take a set of timeseries and add in missing dates as NAs. Preserves source, type, statistic differences in timeseries length. 
intially Fills all regions to be the same length but with trim trailing NAs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$complete(covidTimeseries)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-completeAllRegions"></a>}}
\if{latex}{\out{\hypertarget{method-completeAllRegions}{}}}
\subsection{Method \code{completeAllRegions()}}{
Take a set of timeseries and add in missing dates as NAs. Preserves source, type, statistic differences in timeseries length. 
intially Fills all regions to be the same length but with trim trailing NAs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$completeAllRegions(covidTimeseries)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fillAbsent"></a>}}
\if{latex}{\out{\hypertarget{method-fillAbsent}{}}}
\subsection{Method \code{fillAbsent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$fillAbsent(covidTimeseries, completeDates = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fillAbsentAllRegions"></a>}}
\if{latex}{\out{\hypertarget{method-fillAbsentAllRegions}{}}}
\subsection{Method \code{fillAbsentAllRegions()}}{
Take a set of regional timeseries, finds the full list of regions and the whole timeseries and fills any missing values with zero (if incidence) or the previous value (if cumulative)
explicitly filling in missing values. This is suitable for use for a incidence derived from a line list where some dates and regions will not be observed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$fillAbsentAllRegions(covidTimeseries)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fillAbsentByRegion"></a>}}
\if{latex}{\out{\hypertarget{method-fillAbsentByRegion}{}}}
\subsection{Method \code{fillAbsentByRegion()}}{
Take a set of timeseries and fills any missing values with zero (if incidence) or the previous value (if cumulative)
explicitly filling in zeros for missing values. This is suitable for use for a incidence derived from a line list where some dates will not be
observed. It does not enforce that something has been observed in all regions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$fillAbsentByRegion(covidTimeseries)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fixDatesAndNames"></a>}}
\if{latex}{\out{\hypertarget{method-fixDatesAndNames}{}}}
\subsection{Method \code{fixDatesAndNames()}}{
Take a set of timeseries and fixes any non standard names, truncates time series by truncate days. 
intially Fills all regions to be the same length but with trim trailing NAs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$fixDatesAndNames(covidTimeseries, truncate = NULL)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fixDates"></a>}}
\if{latex}{\out{\hypertarget{method-fixDates}{}}}
\subsection{Method \code{fixDates()}}{
Take a set of timeseries and fixes any non standard names, truncates time series by truncate days. 
intially Fills all regions to be the same length but with trim trailing NAs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$fixDates(covidTimeseries, truncate = NULL)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeseriesFormat dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-timeseriesQA"></a>}}
\if{latex}{\out{\hypertarget{method-timeseriesQA}{}}}
\subsection{Method \code{timeseriesQA()}}{
Check timeseries conforms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$timeseriesQA(covidTimeseries)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeErrors"></a>}}
\if{latex}{\out{\hypertarget{method-describeErrors}{}}}
\subsection{Method \code{describeErrors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$describeErrors(covidTimeseries, valueVar = "value")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe"></a>}}
\if{latex}{\out{\hypertarget{method-describe}{}}}
\subsection{Method \code{describe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$describe(covidTimeseries, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CovidTimeseriesProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
