% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProviderController.R
\name{DataProviderController}
\alias{DataProviderController}
\title{Control data sources}
\description{
Control data sources

Control data sources
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-updateCodes}{\code{DataProviderController$updateCodes()}}
\item \href{#method-updateSPIMDatasets}{\code{DataProviderController$updateSPIMDatasets()}}
\item \href{#method-updateDatasets}{\code{DataProviderController$updateDatasets()}}
\item \href{#method-timeseriesProcessor}{\code{DataProviderController$timeseriesProcessor()}}
\item \href{#method-chessProcessor}{\code{DataProviderController$chessProcessor()}}
\item \href{#method-loadSpimSources}{\code{DataProviderController$loadSpimSources()}}
\item \href{#method-clone}{\code{DataProviderController$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateCodes"></a>}}
\subsection{Method \code{updateCodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProviderController$updateCodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateSPIMDatasets"></a>}}
\subsection{Method \code{updateSPIMDatasets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProviderController$updateSPIMDatasets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateDatasets"></a>}}
\subsection{Method \code{updateDatasets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProviderController$updateDatasets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-timeseriesProcessor"></a>}}
\subsection{Method \code{timeseriesProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProviderController$timeseriesProcessor(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chessProcessor"></a>}}
\subsection{Method \code{chessProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProviderController$chessProcessor(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadSpimSources"></a>}}
\subsection{Method \code{loadSpimSources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProviderController$loadSpimSources(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProviderController$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
