% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetawardProcessingPipeline.R
\name{MetawardProcessingPipeline}
\alias{MetawardProcessingPipeline}
\title{Process metawards output}
\description{
Process metawards output

Process metawards output
}
\section{Super classes}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{\link[ukcovidtools:DataProvider]{ukcovidtools::DataProvider}} -> \code{MetawardProcessingPipeline}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetawardProcessingPipeline$new()}}
\item \href{#method-readSpatialFile}{\code{MetawardProcessingPipeline$readSpatialFile()}}
\item \href{#method-loadAgeTimeMatrix}{\code{MetawardProcessingPipeline$loadAgeTimeMatrix()}}
\item \href{#method-applyAgeTimeMatrix}{\code{MetawardProcessingPipeline$applyAgeTimeMatrix()}}
\item \href{#method-loadDefaultAgeTimeMatrices}{\code{MetawardProcessingPipeline$loadDefaultAgeTimeMatrices()}}
\item \href{#method-clone}{\code{MetawardProcessingPipeline$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="unloadCache">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-unloadCache}{\code{ukcovidtools::PassthroughFilesystemCache$unloadCache()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="ageCatToFactor">}\href{../../ukcovidtools/html/DataProvider.html#method-ageCatToFactor}{\code{ukcovidtools::DataProvider$ageCatToFactor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="ageToAgeCat">}\href{../../ukcovidtools/html/DataProvider.html#method-ageToAgeCat}{\code{ukcovidtools::DataProvider$ageToAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="breakFromCats">}\href{../../ukcovidtools/html/DataProvider.html#method-breakFromCats}{\code{ukcovidtools::DataProvider$breakFromCats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="cutByAge">}\href{../../ukcovidtools/html/DataProvider.html#method-cutByAge}{\code{ukcovidtools::DataProvider$cutByAge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="download">}\href{../../ukcovidtools/html/DataProvider.html#method-download}{\code{ukcovidtools::DataProvider$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadAndUntar">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadAndUntar}{\code{ukcovidtools::DataProvider$downloadAndUntar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadAndUnzip">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadAndUnzip}{\code{ukcovidtools::DataProvider$downloadAndUnzip()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadDaily">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadDaily}{\code{ukcovidtools::DataProvider$downloadDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="labelsFromBreaks">}\href{../../ukcovidtools/html/DataProvider.html#method-labelsFromBreaks}{\code{ukcovidtools::DataProvider$labelsFromBreaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseAgeCat">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseAgeCat}{\code{ukcovidtools::DataProvider$normaliseAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseGender">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseGender}{\code{ukcovidtools::DataProvider$normaliseGender()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetawardProcessingPipeline$new(providerController, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readSpatialFile"></a>}}
\subsection{Method \code{readSpatialFile()}}{
loads a spatial model file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetawardProcessingPipeline$readSpatialFile(fname, ignoreErrors = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{- a filename containing a ward based metaward model .dat file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame containing the simulation results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadAgeTimeMatrix"></a>}}
\subsection{Method \code{loadAgeTimeMatrix()}}{
loads an age time matrix from a csv
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetawardProcessingPipeline$loadAgeTimeMatrix(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{- a url or filename for a csv file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame containing the filter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-applyAgeTimeMatrix"></a>}}
\subsection{Method \code{applyAgeTimeMatrix()}}{
applies a age time matrix to a dataframe containing an incidence, an age category column, and a date column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetawardProcessingPipeline$applyAgeTimeMatrix(
  inputDf,
  matrixDf,
  oldIncidenceVar = "incidence",
  newIncidenceVar = "incidence",
  oldDateVar = "date",
  newDateVar = "date",
  ageCatVar = "ageGroup"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputDf}}{- an input df containing an incidence col, a date col and a ageCategory col this must be grouped}

\item{\code{matrixDf}}{- a dataframe containing a sparse matrix of probabilities where age is labelled}

\item{\code{oldIncidenceVar}}{- the name of the input incidence col}

\item{\code{newIncidenceVar}}{- the name of the output incidence col}

\item{\code{oldDateVar}}{- the name of the input date column}

\item{\code{newDateVar}}{- the name of the output date column}

\item{\code{ageCatVar}}{- the age category variable}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make sure your age categories are the saem as the matrix you are applying
The age time matrix can implement a delay, aggregate over time and or filter by age
}

\subsection{Returns}{
a data frame containing the simulation results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadDefaultAgeTimeMatrices"></a>}}
\subsection{Method \code{loadDefaultAgeTimeMatrices()}}{
loads a ventilator demand matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetawardProcessingPipeline$loadDefaultAgeTimeMatrices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list of matrices
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetawardProcessingPipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
