% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSCapacityProvider.R
\name{NHSCapacityProvider}
\alias{NHSCapacityProvider}
\title{Capacity data from NHS}
\description{
Capacity data from NHS

Capacity data from NHS
}
\section{Super class}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{NHSCapacityProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getHospitals}{\code{NHSCapacityProvider$getHospitals()}}
\item \href{#method-getNightingales}{\code{NHSCapacityProvider$getNightingales()}}
\item \href{#method-getNHSSiteIcuCatchment}{\code{NHSCapacityProvider$getNHSSiteIcuCatchment()}}
\item \href{#method-getNHSTrustIcuCatchment}{\code{NHSCapacityProvider$getNHSTrustIcuCatchment()}}
\item \href{#method-getNHSSiteAcuteCatchment}{\code{NHSCapacityProvider$getNHSSiteAcuteCatchment()}}
\item \href{#method-getNHSTrustAcuteCatchment}{\code{NHSCapacityProvider$getNHSTrustAcuteCatchment()}}
\item \href{#method-getNHSTrustAcuteDemographics}{\code{NHSCapacityProvider$getNHSTrustAcuteDemographics()}}
\item \href{#method-getNHSTrustIcuDemographics}{\code{NHSCapacityProvider$getNHSTrustIcuDemographics()}}
\item \href{#method-getNHSSiteAcuteDemographics}{\code{NHSCapacityProvider$getNHSSiteAcuteDemographics()}}
\item \href{#method-getNHSSiteIcuDemographics}{\code{NHSCapacityProvider$getNHSSiteIcuDemographics()}}
\item \href{#method-clone}{\code{NHSCapacityProvider$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="ageCatToFactor">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-ageCatToFactor}{\code{ukcovidtools::PassthroughFilesystemCache$ageCatToFactor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="ageToAgeCat">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-ageToAgeCat}{\code{ukcovidtools::PassthroughFilesystemCache$ageToAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="breakFromCats">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-breakFromCats}{\code{ukcovidtools::PassthroughFilesystemCache$breakFromCats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="cutByAge">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-cutByAge}{\code{ukcovidtools::PassthroughFilesystemCache$cutByAge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="download">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-download}{\code{ukcovidtools::PassthroughFilesystemCache$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="downloadAndUnzip">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-downloadAndUnzip}{\code{ukcovidtools::PassthroughFilesystemCache$downloadAndUnzip()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="downloadDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-downloadDaily}{\code{ukcovidtools::PassthroughFilesystemCache$downloadDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="importTimeseries">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-importTimeseries}{\code{ukcovidtools::PassthroughFilesystemCache$importTimeseries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="initialize">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-initialize}{\code{ukcovidtools::PassthroughFilesystemCache$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="normaliseGender">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-normaliseGender}{\code{ukcovidtools::PassthroughFilesystemCache$normaliseGender()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHospitals"></a>}}
\subsection{Method \code{getHospitals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getHospitals(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNightingales"></a>}}
\subsection{Method \code{getNightingales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNightingales(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSSiteIcuCatchment"></a>}}
\subsection{Method \code{getNHSSiteIcuCatchment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNHSSiteIcuCatchment(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSTrustIcuCatchment"></a>}}
\subsection{Method \code{getNHSTrustIcuCatchment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNHSTrustIcuCatchment(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSSiteAcuteCatchment"></a>}}
\subsection{Method \code{getNHSSiteAcuteCatchment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNHSSiteAcuteCatchment(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSTrustAcuteCatchment"></a>}}
\subsection{Method \code{getNHSTrustAcuteCatchment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNHSTrustAcuteCatchment(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSTrustAcuteDemographics"></a>}}
\subsection{Method \code{getNHSTrustAcuteDemographics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNHSTrustAcuteDemographics(combineGenders = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSTrustIcuDemographics"></a>}}
\subsection{Method \code{getNHSTrustIcuDemographics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNHSTrustIcuDemographics(combineGenders = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSSiteAcuteDemographics"></a>}}
\subsection{Method \code{getNHSSiteAcuteDemographics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNHSSiteAcuteDemographics(combineGenders = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSSiteIcuDemographics"></a>}}
\subsection{Method \code{getNHSSiteIcuDemographics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$getNHSSiteIcuDemographics(combineGenders = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSCapacityProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
