% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSDatasetProvider.R
\name{NHSDatasetProvider}
\alias{NHSDatasetProvider}
\title{NHS datasets}
\description{
NHS datasets

NHS datasets
}
\section{Super classes}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{\link[ukcovidtools:DataProvider]{ukcovidtools::DataProvider}} -> \code{\link[ukcovidtools:CovidTimeseriesProvider]{ukcovidtools::CovidTimeseriesProvider}} -> \code{NHSDatasetProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NHSDatasetProvider$new()}}
\item \href{#method-getGoogleMobility}{\code{NHSDatasetProvider$getGoogleMobility()}}
\item \href{#method-getSignificantDates}{\code{NHSDatasetProvider$getSignificantDates()}}
\item \href{#method-getPublicOneOneOne}{\code{NHSDatasetProvider$getPublicOneOneOne()}}
\item \href{#method-getTomWhiteCases}{\code{NHSDatasetProvider$getTomWhiteCases()}}
\item \href{#method-getTomWhiteIndicators}{\code{NHSDatasetProvider$getTomWhiteIndicators()}}
\item \href{#method-getNHSDeaths}{\code{NHSDatasetProvider$getNHSDeaths()}}
\item \href{#method-getPHEDashboard}{\code{NHSDatasetProvider$getPHEDashboard()}}
\item \href{#method-getPHEApiNations}{\code{NHSDatasetProvider$getPHEApiNations()}}
\item \href{#method-getPHEApiNHSRegions}{\code{NHSDatasetProvider$getPHEApiNHSRegions()}}
\item \href{#method-getPHEApiNHSTrusts}{\code{NHSDatasetProvider$getPHEApiNHSTrusts()}}
\item \href{#method-getPHEPaginatedData}{\code{NHSDatasetProvider$getPHEPaginatedData()}}
\item \href{#method-getPHEApi}{\code{NHSDatasetProvider$getPHEApi()}}
\item \href{#method-getCLIMB}{\code{NHSDatasetProvider$getCLIMB()}}
\item \href{#method-getCOGUK}{\code{NHSDatasetProvider$getCOGUK()}}
\item \href{#method-getTiers}{\code{NHSDatasetProvider$getTiers()}}
\item \href{#method-getTheFireHose}{\code{NHSDatasetProvider$getTheFireHose()}}
\item \href{#method-clone}{\code{NHSDatasetProvider$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="ageCatToFactor">}\href{../../ukcovidtools/html/DataProvider.html#method-ageCatToFactor}{\code{ukcovidtools::DataProvider$ageCatToFactor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="ageToAgeCat">}\href{../../ukcovidtools/html/DataProvider.html#method-ageToAgeCat}{\code{ukcovidtools::DataProvider$ageToAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="breakFromCats">}\href{../../ukcovidtools/html/DataProvider.html#method-breakFromCats}{\code{ukcovidtools::DataProvider$breakFromCats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="cutByAge">}\href{../../ukcovidtools/html/DataProvider.html#method-cutByAge}{\code{ukcovidtools::DataProvider$cutByAge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="download">}\href{../../ukcovidtools/html/DataProvider.html#method-download}{\code{ukcovidtools::DataProvider$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadAndUntar">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadAndUntar}{\code{ukcovidtools::DataProvider$downloadAndUntar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadAndUnzip">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadAndUnzip}{\code{ukcovidtools::DataProvider$downloadAndUnzip()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="downloadDaily">}\href{../../ukcovidtools/html/DataProvider.html#method-downloadDaily}{\code{ukcovidtools::DataProvider$downloadDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="labelsFromBreaks">}\href{../../ukcovidtools/html/DataProvider.html#method-labelsFromBreaks}{\code{ukcovidtools::DataProvider$labelsFromBreaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseAgeCat">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseAgeCat}{\code{ukcovidtools::DataProvider$normaliseAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="DataProvider" data-id="normaliseGender">}\href{../../ukcovidtools/html/DataProvider.html#method-normaliseGender}{\code{ukcovidtools::DataProvider$normaliseGender()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="complete">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-complete}{\code{ukcovidtools::CovidTimeseriesProvider$complete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="completeAllRegions">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-completeAllRegions}{\code{ukcovidtools::CovidTimeseriesProvider$completeAllRegions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="describe">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-describe}{\code{ukcovidtools::CovidTimeseriesProvider$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="describeErrors">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-describeErrors}{\code{ukcovidtools::CovidTimeseriesProvider$describeErrors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="fillAbsent">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-fillAbsent}{\code{ukcovidtools::CovidTimeseriesProvider$fillAbsent()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="fillAbsentAllRegions">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-fillAbsentAllRegions}{\code{ukcovidtools::CovidTimeseriesProvider$fillAbsentAllRegions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="fillAbsentByRegion">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-fillAbsentByRegion}{\code{ukcovidtools::CovidTimeseriesProvider$fillAbsentByRegion()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="fixDatesAndNames">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-fixDatesAndNames}{\code{ukcovidtools::CovidTimeseriesProvider$fixDatesAndNames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="setTruncation">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-setTruncation}{\code{ukcovidtools::CovidTimeseriesProvider$setTruncation()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="timeseriesQA">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-timeseriesQA}{\code{ukcovidtools::CovidTimeseriesProvider$timeseriesQA()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="CovidTimeseriesProvider" data-id="trimNAs">}\href{../../ukcovidtools/html/CovidTimeseriesProvider.html#method-trimNAs}{\code{ukcovidtools::CovidTimeseriesProvider$trimNAs()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$new(providerController, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGoogleMobility"></a>}}
\subsection{Method \code{getGoogleMobility()}}{
Load google mobility file aligned with LAD codes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getGoogleMobility(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
google mobility 0 custom format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSignificantDates"></a>}}
\subsection{Method \code{getSignificantDates()}}{
Load curated significant dates file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getSignificantDates(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
public 111 data including pathways and online, pre and post april. Codes are CCG20 codes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPublicOneOneOne"></a>}}
\subsection{Method \code{getPublicOneOneOne()}}{
Load public 111 summary file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPublicOneOneOne(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
public 111 data including pathways and online, pre and post april. Codes are CCG20 codes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTomWhiteCases"></a>}}
\subsection{Method \code{getTomWhiteCases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getTomWhiteCases(truncate = 4, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTomWhiteIndicators"></a>}}
\subsection{Method \code{getTomWhiteIndicators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getTomWhiteIndicators(truncate = 4, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNHSDeaths"></a>}}
\subsection{Method \code{getNHSDeaths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getNHSDeaths(truncate = 7, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPHEDashboard"></a>}}
\subsection{Method \code{getPHEDashboard()}}{
Get UK outbreak timeseries data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPHEDashboard(truncate = 4, ...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a covidTimeserisFormat data frame with several timeseries in it
Extracts paginated data by requesting all of the pages
and combining the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPHEApiNations"></a>}}
\subsection{Method \code{getPHEApiNations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPHEApiNations(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPHEApiNHSRegions"></a>}}
\subsection{Method \code{getPHEApiNHSRegions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPHEApiNHSRegions(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPHEApiNHSTrusts"></a>}}
\subsection{Method \code{getPHEApiNHSTrusts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPHEApiNHSTrusts(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPHEPaginatedData"></a>}}
\subsection{Method \code{getPHEPaginatedData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPHEPaginatedData(filters, structure)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filters}}{API filters. See the API documentations for 
additional information.}

\item{\code{structure}}{Structure parameter. See the API documentations 
for additional information.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list      Comprehensive list of dictionaries containing all 
                  the data for the given ``filter`` and ``structure`.`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPHEApi"></a>}}
\subsection{Method \code{getPHEApi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getPHEApi(
  areaType = "nation",
  areaName = NULL,
  areaCode = NULL,
  truncate = 4,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCLIMB"></a>}}
\subsection{Method \code{getCLIMB()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getCLIMB(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCOGUK"></a>}}
\subsection{Method \code{getCOGUK()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getCOGUK(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTiers"></a>}}
\subsection{Method \code{getTiers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getTiers(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTheFireHose"></a>}}
\subsection{Method \code{getTheFireHose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$getTheFireHose(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NHSDatasetProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
