% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileProvider.R
\name{ReadOnlyFileProvider}
\alias{ReadOnlyFileProvider}
\title{Read only file providers}
\description{
Read only file providers

Read only file providers
}
\section{Super class}{
\code{\link[ukcovidtools:AbstractFileProvider]{ukcovidtools::AbstractFileProvider}} -> \code{ReadOnlyFileProvider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getFile}{\code{ReadOnlyFileProvider$getFile()}}
\item \href{#method-getZip}{\code{ReadOnlyFileProvider$getZip()}}
\item \href{#method-processFiles}{\code{ReadOnlyFileProvider$processFiles()}}
\item \href{#method-print}{\code{ReadOnlyFileProvider$print()}}
\item \href{#method-clone}{\code{ReadOnlyFileProvider$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="initialize">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-initialize}{\code{ukcovidtools::AbstractFileProvider$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="listAllFiles">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-listAllFiles}{\code{ukcovidtools::AbstractFileProvider$listAllFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="listFiles">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-listFiles}{\code{ukcovidtools::AbstractFileProvider$listFiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="AbstractFileProvider" data-id="relativePath">}\href{../../ukcovidtools/html/AbstractFileProvider.html#method-relativePath}{\code{ukcovidtools::AbstractFileProvider$relativePath()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFile"></a>}}
\if{latex}{\out{\hypertarget{method-getFile}{}}}
\subsection{Method \code{getFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadOnlyFileProvider$getFile(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getZip"></a>}}
\if{latex}{\out{\hypertarget{method-getZip}{}}}
\subsection{Method \code{getZip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadOnlyFileProvider$getZip(filename, asFile = FALSE, unzipDir = tempdir())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-processFiles"></a>}}
\if{latex}{\out{\hypertarget{method-processFiles}{}}}
\subsection{Method \code{processFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadOnlyFileProvider$processFiles(func, paths)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadOnlyFileProvider$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadOnlyFileProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
