% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalProcessingPipeline.R
\name{SurvivalProcessingPipeline}
\alias{SurvivalProcessingPipeline}
\title{Survival data procesing}
\description{
Survival data procesing

Survival data procesing
}
\section{Super class}{
\code{\link[ukcovidtools:PassthroughFilesystemCache]{ukcovidtools::PassthroughFilesystemCache}} -> \code{SurvivalProcessingPipeline}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-generateSurvivalData}{\code{SurvivalProcessingPipeline$generateSurvivalData()}}
\item \href{#method-getMaxDate}{\code{SurvivalProcessingPipeline$getMaxDate()}}
\item \href{#method-transitionMatrix}{\code{SurvivalProcessingPipeline$transitionMatrix()}}
\item \href{#method-generateMultistateSurvivalData}{\code{SurvivalProcessingPipeline$generateMultistateSurvivalData()}}
\item \href{#method-standardModels}{\code{SurvivalProcessingPipeline$standardModels()}}
\item \href{#method-fitModelsByAge}{\code{SurvivalProcessingPipeline$fitModelsByAge()}}
\item \href{#method-doPdf}{\code{SurvivalProcessingPipeline$doPdf()}}
\item \href{#method-doCdf}{\code{SurvivalProcessingPipeline$doCdf()}}
\item \href{#method-doMedian}{\code{SurvivalProcessingPipeline$doMedian()}}
\item \href{#method-doEstimate}{\code{SurvivalProcessingPipeline$doEstimate()}}
\item \href{#method-createSurvivalSurfaces}{\code{SurvivalProcessingPipeline$createSurvivalSurfaces()}}
\item \href{#method-plotSurvivalSurface}{\code{SurvivalProcessingPipeline$plotSurvivalSurface()}}
\item \href{#method-plotIncidenceSurface}{\code{SurvivalProcessingPipeline$plotIncidenceSurface()}}
\item \href{#method-plotProbabilityMatrix}{\code{SurvivalProcessingPipeline$plotProbabilityMatrix()}}
\item \href{#method-clone}{\code{SurvivalProcessingPipeline$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="ageCatToFactor">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-ageCatToFactor}{\code{ukcovidtools::PassthroughFilesystemCache$ageCatToFactor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="ageToAgeCat">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-ageToAgeCat}{\code{ukcovidtools::PassthroughFilesystemCache$ageToAgeCat()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="breakFromCats">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-breakFromCats}{\code{ukcovidtools::PassthroughFilesystemCache$breakFromCats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="cutByAge">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-cutByAge}{\code{ukcovidtools::PassthroughFilesystemCache$cutByAge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="download">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-download}{\code{ukcovidtools::PassthroughFilesystemCache$download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="downloadAndUnzip">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-downloadAndUnzip}{\code{ukcovidtools::PassthroughFilesystemCache$downloadAndUnzip()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="downloadDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-downloadDaily}{\code{ukcovidtools::PassthroughFilesystemCache$downloadDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getDaily">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getDaily}{\code{ukcovidtools::PassthroughFilesystemCache$getDaily()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getHashCached">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getHashCached}{\code{ukcovidtools::PassthroughFilesystemCache$getHashCached()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="getSaved">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-getSaved}{\code{ukcovidtools::PassthroughFilesystemCache$getSaved()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="importTimeseries">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-importTimeseries}{\code{ukcovidtools::PassthroughFilesystemCache$importTimeseries()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="initialize">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-initialize}{\code{ukcovidtools::PassthroughFilesystemCache$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ukcovidtools" data-topic="PassthroughFilesystemCache" data-id="normaliseGender">}\href{../../ukcovidtools/html/PassthroughFilesystemCache.html#method-normaliseGender}{\code{ukcovidtools::PassthroughFilesystemCache$normaliseGender()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateSurvivalData"></a>}}
\subsection{Method \code{generateSurvivalData()}}{
generate survival data and right censored data from
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$generateSurvivalData(
  df,
  idVar,
  startDateVar,
  endDateExpr,
  statusExpr = if_else(is.na({     {         endDateExpr     } }), 0, 1),
  ageVar = "age",
  censoredDateExpr = max({     {         endDateExpr     } }, na.rm = TRUE),
  ageBreaks = c(-Inf, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80,
    Inf),
  ageLabels = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+"),
  ageReferenceCat = NA,
  statusLabels = c("censored", "died")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the ff100 file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data set in survival format including status, time, left, right, ageCat columns
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMaxDate"></a>}}
\subsection{Method \code{getMaxDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$getMaxDate(df, dateCols)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transitionMatrix"></a>}}
\subsection{Method \code{transitionMatrix()}}{
generate transition matrix with enquoted tidy column sepecifications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$transitionMatrix(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{a set of mappings}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a transistion matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateMultistateSurvivalData"></a>}}
\subsection{Method \code{generateMultistateSurvivalData()}}{
generate survival data and right censored data from a multistate model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$generateMultistateSurvivalData(
  df,
  idVar,
  transMatrix,
  ageVar = "age",
  censoringDateExpr = getMaxDate(df, dateCols),
  startDateCol = names(allowableTransitions)[1],
  ageBreaks = c(-Inf, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80,
    Inf),
  ageLabels = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+"),
  ageReferenceCat = NA,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{- path to the ff100 file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data set in survival format including status, time, left, right, ageCat columns
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardModels"></a>}}
\subsection{Method \code{standardModels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$standardModels(
  items = c("weibull", "gamma", "lognorm")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fitModelsByAge"></a>}}
\subsection{Method \code{fitModelsByAge()}}{
for each ageCat fit a set of models and return a data fram with the parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$fitModelsByAge(
  survivalDf,
  ageVar = "ageCat",
  models = standardModels(),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{survivalDf}}{- a data frame containing an ageCat, and a left and right column with min and max times of individual events, If event is censored then right is NA}

\item{\code{models}}{-  a list of models that you want to fit - e.g. lnorm, norm, etc...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an datafram of ageCat, model, param(eterName), value, low_ci, high_ci
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doPdf"></a>}}
\subsection{Method \code{doPdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$doPdf(model, paramNames, params, timepoints, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doCdf"></a>}}
\subsection{Method \code{doCdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$doCdf(model, paramNames, params, timepoints, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doMedian"></a>}}
\subsection{Method \code{doMedian()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$doMedian(model, paramNames, params, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doEstimate"></a>}}
\subsection{Method \code{doEstimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$doEstimate(d, model1, timepoints, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createSurvivalSurfaces"></a>}}
\subsection{Method \code{createSurvivalSurfaces()}}{
for a set of model parameters, calculate the distributions and median
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$createSurvivalSurfaces(
  modelParameterDf,
  days = 30,
  timepoints = seq(0.1, days, length.out = 100)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelParameterDf}}{- a data frame set of parameters for a range of models as param, value columns}

\item{\code{days}}{- the time frame of the prediction}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an dataframe of ageCat and time varying surfaces, including pdfs, cdfs and medians, with high and low estimates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSurvivalSurface"></a>}}
\subsection{Method \code{plotSurvivalSurface()}}{
plot a survival surface as a CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$plotSurvivalSurface(surfacesDf, distr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surfacesDf}}{- an dataframe of ageCat and time varying surfaces, including pdfs, cdfs and medians, with high and low estimates}

\item{\code{distr}}{- a single distribution to plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a plot of the survival surface (1-cdf)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotIncidenceSurface"></a>}}
\subsection{Method \code{plotIncidenceSurface()}}{
plot a survival surface as a PDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$plotIncidenceSurface(surfacesDf, distr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surfacesDf}}{- an dataframe of ageCat and time varying surfaces, including pdfs, cdfs and medians, with high and low estimates}

\item{\code{distr}}{- a single distribution to plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a plot of the survival surface as pdf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotProbabilityMatrix"></a>}}
\subsection{Method \code{plotProbabilityMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$plotProbabilityMatrix(
  surfacesDf,
  distr,
  pExpr = "p"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalProcessingPipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
